Added data verification
Updated the html and js for the post creation webpage to add data verification. I made sure the user can't exceed character limits, that they can't add empty values, that they can't upload giant pictures, and more. I also uploaded a 5mb+ pic to the github so that people who clone it to download will have something to test with.